requirements.txt
streamlit
langchain
langchain-community
langchain-core
langchain-text-splitters
langchain-chroma
langchain-groq
langchain-huggingface
chromadb
pypdf
python-dotenv


